test : Lotto match test
